Telegram_bot/requirements.txt
python-dotenv
outline-api
python-telegram-bot[job-queue]
requests
asyncpg
sqlalchemy[asyncio]
aiosqlite